Preallocate `n_iter_` array
Before starting the `n_targets` `for`-loop, preallocate an array for
`n_iter_` to store each iteration length into as we know how many values
we need and of what type. This is a pretty mild improvement by itself,
but it should make it a bit easier to Cythonize the code and parallelize
this loop.